[GUI] Minor changes

+++ src/main/java/Model/Ship.java
@@ -3,12 +3,12 @@ package Model;
 public class Ship {
 
     public String name;
-    public float tonnage;
-    public float maxVelocity;
-    public float fuelAmount;
-    public float fuelConsumptionRate;
+    public double tonnage;
+    public double maxVelocity;
+    public double fuelAmount;
+    public double fuelConsumptionRate;
 
-    public Ship(String name, float tonnage, float maxVelocity, float fuelAmount, float fuelConsumptionRate) {
+    public Ship(String name, double tonnage, double maxVelocity, double fuelAmount, double fuelConsumptionRate) {
         this.name = name;
         this.tonnage = tonnage;
         this.maxVelocity = maxVelocity;
